Fix compilation errors when \link defined inside \section titles
This was a nightmare to pin down. There were a few inconsistent,
apparently inoccuous errors, which warned against this, but they
appeared spurrious as tthey claimed that the \links were
multiply-defined.

This was the source of the problems with compilation, where links
would not show up correctly unless the document was recompiled.
It seemed to be an issues with cross-referencing, i.e. it takes
several passes for output to 'settle' as regards to somewhat-nested
links.

This commit fixes this behaviour and the document now consistently
compiles with latexmk(1), rubber(1), and Overleaf.
There do seem to be a number of passes required, due to the hyperref
package, and potentially-nested links.